requirements.txt
Django==2.1
djangorestframework==3.10.3
django-filter==2.2.0
Markdown==3.1.1
psycopg2==2.8.3
pytest==5.1.2
pytest-django==3.5.1
pytest-cov==2.7.1
python-decouple==3.1
dj-database-url==0.5
dj-static==0.0.6
gunicorn==19.9
